+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to
 - The canonical format (`-Tcanon`) produces output that is stable across
   multiple invocations. #2193
 
+- GDKPixbuf JPEG output is no longer empty with GDKPixbuf 2.44.
+
 ## [16.1.0] – 2026-09-03
 
 ### Added